social media API/requirements.txt
absl-py==2.3.1
africastalking==1.2.9
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==5.5.0
asgiref==3.9.1
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Django==5.2.4
django-csp==4.0
django-taggit==6.1.0
djangorestframework==3.16.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
frozenlist==1.7.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.73.1
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==3.10.0
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
multidict==6.6.3
namex==0.1.0
narwhals==1.47.1
numpy==2.1.3
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.1
Pillow==11.3.0
propcache==0.3.2
protobuf==5.29.5
pyarrow==21.0.0
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
schema==0.7.7
six==1.17.0
smmap==5.0.2
soupsieve==2.7
SQLAlchemy==2.0.43
sqlparse==0.5.3
streamlit==1.47.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
twilio==9.6.5
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.1
